+++ src/helpers/types.ts
@@ -9,8 +9,7 @@ import {
 } from "graphql";
 
 import { TypeOptions } from "../decorators/types";
-import { GraphQLTimestamp } from "../scalars/timestamp";
-import { GraphQLISODateTime } from "../scalars/isodate";
+import { GraphQLISODateTime } from "../scalars";
 import { BuildContext } from "../schema/build-context";
 import { WrongNullableListOptionError } from "../errors";
 
@@ -31,7 +30,7 @@ export function convertTypeIfScalar(type: any): GraphQLScalarType | undefined {
     case Number:
       return GraphQLFloat;
     case Date:
-      return BuildContext.dateScalarMode === "isoDate" ? GraphQLISODateTime : GraphQLTimestamp;
+      return GraphQLISODateTime;
     default:
       return undefined;
   }